+++ src/kernel_evm/kernel/src/tick_model.rs
@@ -38,7 +38,7 @@ const SAFETY_MARGIN: u64 = 2_000_000_000;
 /// estimate value using benchmarks
 const INITIALISATION_OVERHEAD: u64 = 1_000_000_000;
 /// see [maximum_reboots_per_input]
-const _MAXIMUM_REBOOTS: u64 = 1_000;
+pub const _MAXIMUM_NUMBER_OF_REBOOTS: u32 = 1_000;
 /// The minimum amount of gas for an ethereum transaction.
 const BASE_GAS: u64 = Config::london().gas_transaction_call;
 
@@ -76,7 +76,7 @@ pub fn estimate_would_overflow(estimated_ticks: u64, transaction: &Transaction) 
 
 /// Initial amount for the tick accumulator, corresponding to the overhead of a
 /// block
-pub fn block_overhead_ticks() -> u64 {
+pub fn top_level_overhead_ticks() -> u64 {
     INITIALISATION_OVERHEAD
 }